TST: move fastparquet BytesIO regression test to test_parquet.py
